introduce ``HasParent`` as Trait for entities - move page logic of ``HasParent`` to its on trait and did minor refactoring - add ``HasParent`` to Database and Block entity
