+++ src/Entities/Blocks/Block.php
@@ -4,6 +4,7 @@ namespace FiveamCode\LaravelNotionApi\Entities\Blocks;
 
 use FiveamCode\LaravelNotionApi\Entities\Entity;
 use FiveamCode\LaravelNotionApi\Exceptions\HandlingException;
+use FiveamCode\LaravelNotionApi\Traits\HasParent;
 use FiveamCode\LaravelNotionApi\Traits\HasTimestamps;
 use Illuminate\Support\Arr;
 
@@ -12,7 +13,7 @@ use Illuminate\Support\Arr;
  */
 class Block extends Entity
 {
-    use HasTimestamps;
+    use HasTimestamps, HasParent;
 
     /**
      * @var string
@@ -61,6 +62,7 @@ class Block extends Entity
         $this->fillType();
         $this->fillRawContent();
         $this->fillHasChildren();
+        $this->fillParentProperties();
         $this->fillTimestampableProperties();
     }
